Make MockChatLog reusable for other integrations
