+++ tests/components/conversation/__init__.py
@@ -2,7 +2,11 @@
 
 from __future__ import annotations
 
+from dataclasses import dataclass, field
 from typing import Literal
+from unittest.mock import patch
+
+import pytest
 
 from homeassistant.components import conversation
 from homeassistant.components.conversation.models import (
@@ -14,7 +18,7 @@ from homeassistant.components.homeassistant.exposed_entities import (
     async_expose_entity,
 )
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import intent
+from homeassistant.helpers import chat_session, intent
 
 
 class MockAgent(conversation.AbstractConversationAgent):
@@ -42,6 +46,53 @@ class MockAgent(conversation.AbstractConversationAgent):
         return ConversationResult(
             response=response, conversation_id=user_input.conversation_id
         )
+
+
+@pytest.fixture
+async def mock_chat_log(hass: HomeAssistant) -> conversation.MockChatLog:
+    """Return mock chat logs."""
+    # pylint: disable-next=contextmanager-generator-missing-cleanup
+    with (
+        patch(
+            "homeassistant.components.conversation.chat_log.ChatLog",
+            MockChatLog,
+        ),
+        chat_session.async_get_chat_session(hass, "mock-conversation-id") as session,
+        conversation.async_get_chat_log(hass, session) as chat_log,
+    ):
+        yield chat_log
+
+
+@dataclass
+class MockChatLog(conversation.ChatLog):
+    """Mock chat log."""
+
+    _mock_tool_results: dict = field(default_factory=dict)
+
+    def mock_tool_results(self, results: dict) -> None:
+        """Set tool results."""
+        self._mock_tool_results = results
+
+    @property
+    def llm_api(self):
+        """Return LLM API."""
+        return self._llm_api
+
+    @llm_api.setter
+    def llm_api(self, value):
+        """Set LLM API."""
+        self._llm_api = value
+
+        if not value:
+            return
+
+        async def async_call_tool(tool_input):
+            """Call tool."""
+            if tool_input.id not in self._mock_tool_results:
+                raise ValueError(f"Tool {tool_input.id} not found")
+            return self._mock_tool_results[tool_input.id]
+
+        self._llm_api.async_call_tool = async_call_tool
 
 
 def expose_new(hass: HomeAssistant, expose_new: bool) -> None:
